# src/backend/docs/specs.openapi.yml
openapi: 3.0.1
info:
  title: Anekdotas MindGame
  description: "Quiz application to advance and/or evaluate players' knowledge in different categories and topics."
  contact:
    email: borysej90@gmail.com
  version: alpha
...
paths:
...
  /users/stats:
    get:
      tags:
        - users
      summary: Gets User's statistics
      operationId: getUserStats
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalTimeSpent:
                    type: integer
                  correctAnswers:
                    type: integer
                  correctAnswersPercentage:
                    type: number
                    format: float
                  longestStreak:
                    type: integer
                  longestStreakTopic:
                    type: string
                  averageGameTime:
                    type: integer
                  topicsCreated:
                    type: integer
                  topicsRated:
                    type: integer
                  topicsPlayed:
                    type: integer
        401:
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    oAuth2:
      type: oauth2
      flows:
        password:
          tokenUrl: /auth/login
          scopes: {}
  responses:
...
    Unauthorized:
      description: User does not have permission to access the endpoint
...